% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_pickr.R
\name{color_pickr}
\alias{color_pickr}
\title{Color Picker}
\usage{
color_pickr(n, palette)
}
\arguments{
\item{n}{the number of colors needed for the palette}

\item{palette}{string name of the palette to draw colors from, "Jalama", "Refugio", "Rincon", or "El Capitan"}
}
\value{
This function returns a \code{vector} of hex color codes
}
\description{
This function returns a palette of colors optimized for the requested number of colors
}
\author{
Bjorn Kallerud
}
