% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themer.R
\name{themer}
\alias{themer}
\title{Theme for ggplot2}
\usage{
themer(font_size = 12, box = T, flipped = F)
}
\arguments{
\item{font_size}{adjust plot font size}

\item{box}{determines whether or not there is a box around the plot}

\item{flipped}{set to TRUE if you are using coord_flip}
}
\value{
This function returns a \code{vector} of hex color codes
}
\description{
This function automates and standardizes a ggplot2 theme
}
\author{
Bjorn Kallerud
}
