% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themer.R
\name{themer}
\alias{themer}
\title{Theme for ggplot2}
\usage{
themer(font_size = 12, box = T)
}
\arguments{
\item{n}{the number of colors needed for the palette}

\item{palette}{string name of the palette to draw colors from, "Jalama", "Refugio", "Rincon", or "El Capitan"}
}
\value{
This function returns a \code{vector} of hex color codes
}
\description{
This function automates and standardizes a ggplot2 theme
}
\author{
Bjorn Kallerud
}
